build: add most of bugprone check in tidy

+++ src/engine/src/scheduler/FixedTimeUpdate.cpp
@@ -15,7 +15,7 @@ void Engine::Scheduler::FixedTimeUpdate::RunSystems()
 {
     _bufferedTime += this->_core.GetResource<Engine::Resource::Time>()._elapsedTime;
     auto ticks = static_cast<unsigned int>(_bufferedTime / _tickRate);
-    _bufferedTime -= ticks * _tickRate;
+    _bufferedTime -= static_cast<float>(ticks) * _tickRate;
 
     for (unsigned int i = 0; i < ticks; i++)
     {
